more int fixes

+++ include/controllers/default/cockpit2/api/staff/payinfo.php
@@ -8,12 +8,12 @@ class Controller_api_staff_payinfo extends Crunchbutton_Controller_RestAccount {
 
 		if( $hasPermission ){
 			if( c::getPagePiece( 3 ) ){
-				$admin = Admin::o( c::getPagePiece( 3 ) );
+				$admin = Admin::o((int)c::getPagePiece( 3 ) );
 				if (!$admin->id_admin) {
 					$admin = Admin::login(c::getPagePiece(3), true);
 				}
 			} else {
-				$admin = Admin::o( $this->request()[ 'id_admin' ] );
+				$admin = Admin::o((int)$this->request()[ 'id_admin' ] );
 				if (!$admin->id_admin) {
 					$admin = Admin::login($this->request()['id_admin'], true);
 				}
